app/requirements.txt
google-cloud-texttospeech==2.14.1
google-api-python-client==2.81.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-login==0.6.2
mecab-python3==1.0.6
mysql-connector-python==8.0.32
pytest==7.2.2
gunicorn==20.1.0
Werkzeug==2.2.2
python-dotenv
pymysql>=1.1.0
cryptography